backend/requirements.render.txt
# Smart Rental Tracker - Production Requirements for Render
# Optimized for deployment with pinned versions

# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database ORM and SQLite
sqlalchemy==2.0.23
alembic==1.12.1

# Data Validation and Serialization
pydantic==2.5.0

# Email and Notification System
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0

# ML Dependencies for Analytics and Forecasting
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.24.4
matplotlib==3.7.3
seaborn==0.13.0
joblib==1.3.2

# Data Processing and File Handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# Production Server
gunicorn==21.2.0

# Environment and Configuration
python-multipart==0.0.6
